Update hardware-wallet-recommendations.md
I added the word, 'than', (note: NOT then), before $100 on the second line and removed commas wherever they appeared incorrectly before the word, 'and'. I just did it in the context of highlighting the word but in normal sentences, 'and' is never directly preceded by a comma as it is used in place of what would otherwise be the last comma in a list, the commas only being there to avoid saying, 'and' repeatedly as links in the chain, as it were. The same goes for, 'or'.

I do appreciate that language evolves over time and that 'writing like you talk' is a welcome trend. However, whilst some rules are flexible, basic sentence structure should not be one of them, as it undermines the critical thinking skills necessary to navigate an increasingly corrupt world. Corrupting the language was the first step in that process but I digress.